añadido agregar nuevo compuesto, ver compuesto, agregar-sacar compuestos

+++ src/main/resources/db.migration/V1__create-table-compuestos.sql
@@ -2,7 +2,7 @@ CREATE TABLE Compuestos(
     id INT PRIMARY KEY AUTO_INCREMENT,
     nombreCompuesto VARCHAR(255) NOT NULL,
     organico BOOLEAN NOT NULL,
-    peso INT NOT NULL,
+    peso double NOT NULL,
     explosivo BOOLEAN NOT NULL,
     inflamable BOOLEAN NOT NULL,
     carburante BOOLEAN NOT NULL,
